CVSL-3319 This is a WIP, the basic idea of storey update after review - removed spring bean!
